python_core/requirements.txt
pytesseract
opencv-python-headless
pdf2image
Pillow